models/f5.rb: require f5-icontrol
Add require 'f5-icontrol' to the F5 model.

Previously, the F5 controller required 'f5-icontrol', but that was changed
in commit da699d4cc2a3a002de2a5639655e7ff0be5450bb so that the F5
controller could be used with other load-balancers.  However, we do need
the require statement for the model.