daemon.rb: Use STOMP 1.1 and nack on errors
initialize: Specify the appropriate headers to force the client to use
STOMP 1.1 (rather than the default of STOMP 1.0).

initialize: Use :ack => 'client-individual' (instead of :ack => 'client') so
that each message must be acknowledged.

listen: Check whether the received message is non-nil and has a message-id
header and skip it if not.

listen: Specify the 'subscription' when acknowledging a message, for STOMP
1.1 compatibility.

listen: Rescue Psych::SyntaxError exceptions.

listen: Rescue LBControllerException and LBModelException and nack the
message from ActiveMQ if one of these exceptions is raised.

listen: Use an ensure clause to update the daemon if necessary, in order to
simplify the logic.