daemon.rb: Create monitor iff name is unique
create_application: Instead of checking whether the monitor already exists,
check whether the template will be unique to the application (as indicated
by the presence of the %a and %n expandos in the name).

delete_application: Likewise.

The intention is to reduce coupling between the daemon and the controller
and not require having access to the list of monitors in the daemon.