daemon.rb: Fix log message for deleting pool
delete_application: Change the log message from "Deleting empty pool" to
"Deleting pool."

After commit 1247362f28259fe7ceaf47d26446fe1aa11051d5, delete_application
no longer checks that the pool is empty, instead relying on the
controller's or model's raising an exception should the pool in fact not be
empty when the daemon tries to delete it.  This commit updates the log
message so as to be less confusing in this contingency.