openshift.ks: Start services; run oo-diagnostics
Add restart_services, a new function that starts or restarts required
services based on the setting for install_components.

Add run_diagnostics, a new function that invokes oo-diagnostics -v.

Modify do_all_actions to invoke restart_services and run_diagnostics
and not to print instructions to reboot or run oo-diagnostics.