Introduce REST iControl model for F5 BIG-IP
Introduce a REST iControl model for F5 BIG-IP, and use it instead of the
SOAP iControl model.

The REST iControl interface has the advantages that it does not need the
f5-icontrol rubygem and that it allows us to manipulate local-traffic
policies.

BIG-IP 11.4 deprecates the HTTP class profile feature and introduces the
local-traffic policies feature in its stead, so we need to use the
latter now, which means we need to use REST.

Update README.md accordingly.

Update README.md for change from SOAP to REST