controllers/f5_batched.rb: Fix lazy loading
Change the F5-batched controller code to refer to pools, members, and
routes rather than to @pools, @members, and @routes as the former will
lazily initialize the latter, but the latter will be nil and will not
lazily initialize when referenced directly.

This is essentially the same change that commit
f0b0655714bd7e88a584fb719aa38d9477eec260 made to the LBaaS controller.