openshift.ks: Install time package on broker
install_broker_pkgs: Install the time package so that the external time
command will be available for use in configure_selinux_policy_on_broker.

Because we use the time command in the right-hand side of a pipeline, we
must have the external command available rather than relying on Bash's
builtin time command; Bash does not allow builtins to be used on the
right-hand side of a pipeline.

This commit fixes bug 1158688.